add input time exit amout

+++ src/presentation/input/pedestrian/PedestrianInput.java
@@ -1,5 +1,6 @@
 package presentation.input.pedestrian;
 
+import javax.swing.*;
 import java.awt.*;
 import java.util.ArrayList;
 
@@ -7,6 +8,6 @@ public class PedestrianInput {
     public Point pedestrianEntry;
     public Point pedestrianExit;
     public int amountOfPedestrian = 0;
-    public int timeIn = 0;
+    public long timeIn = 0;
     public ArrayList<Point> way;
 }
